+++ versions.tf
@@ -7,7 +7,7 @@ terraform {
     hcloud = {
       # Documentation; https://registry.terraform.io/providers/hetznercloud/hcloud
       source  = "hetznercloud/hcloud"
-      version = "~> 1.59.0"
+      version = "~> 1.60.0"
     }
     http = {
       source  = "hashicorp/http"
@@ -15,7 +15,7 @@ terraform {
     }
     local = {
       source  = "hashicorp/local"
-      version = "~> 2.6.0"
+      version = "~> 2.8.0"
     }
     random = {
       source  = "hashicorp/random"